new strategy to generate roles
